+++ src/test/java/com/umc5th/muffler/domain/expense/service/ExpenseServiceTest.java
@@ -1,10 +1,5 @@
 package com.umc5th.muffler.domain.expense.service;
 
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.when;
-import static org.mockito.MockitoAnnotations.openMocks;
-
-import com.umc5th.muffler.domain.category.repository.CategoryRepository;
 import com.umc5th.muffler.domain.expense.dto.DailyExpenseDetailsResponse;
 import com.umc5th.muffler.domain.expense.dto.WeeklyExpenseDetailsResponse;
 import com.umc5th.muffler.domain.expense.repository.ExpenseRepository;
@@ -12,7 +7,7 @@ import com.umc5th.muffler.domain.member.repository.MemberRepository;
 import com.umc5th.muffler.entity.Expense;
 import com.umc5th.muffler.entity.Member;
 import com.umc5th.muffler.fixture.ExpenseFixture;
-import com.umc5th.muffler.fixture.MemberEntityFixture;
+import com.umc5th.muffler.fixture.MemberFixture;
 import com.umc5th.muffler.global.response.exception.MemberException;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -52,7 +47,7 @@ class ExpenseServiceTest {
         Pageable pageable = PageRequest.of(0, pageSize);
         Long memberId = 1L;
 
-        Member mockMember = MemberEntityFixture.create();
+        Member mockMember = MemberFixture.create();
 
         List<Expense> expenses = ExpenseFixture.createList(10, testDate);
         Slice<Expense> expenseSlice = new SliceImpl<>(expenses, pageable, false);
@@ -96,7 +91,7 @@ class ExpenseServiceTest {
         LocalDate date = LocalDate.of(2024, 1, 1);
         LocalDate startDate = date.with(TemporalAdjusters.previousOrSame(DayOfWeek.MONDAY));
         LocalDate endDate = date.with(TemporalAdjusters.nextOrSame(DayOfWeek.SUNDAY));
-        Member mockMember = MemberEntityFixture.create();
+        Member mockMember = MemberFixture.create();
 
         List<Expense> expenses = ExpenseFixture.createList(20, startDate);
         Slice<Expense> expenseSlice = new SliceImpl<>(expenses, pageable, true);
